+++ sdk/monitor/azure-monitor-opentelemetry-exporter/azure/monitor/opentelemetry/exporter/statsbeat/_state.py
@@ -5,8 +5,7 @@ import threading
 from typing import Dict, Union
 
 from azure.monitor.opentelemetry.exporter._constants import (
-    _APPLICATIONINSIGHTS_STATSBEAT_DISABLED_ALL,
-    _APPLICATIONINSIGHTS_SDKSTATS_ENABLED_PREVIEW
+    _APPLICATIONINSIGHTS_STATSBEAT_DISABLED_ALL
 )
 
 _REQUESTS_MAP: Dict[str, Union[int, Dict[int, int]]] = {}
@@ -21,24 +20,6 @@ _STATSBEAT_STATE = {
 }
 _STATSBEAT_STATE_LOCK = threading.Lock()
 _STATSBEAT_FAILURE_COUNT_THRESHOLD = 3
-
-
-_CUSTOMER_SDKSTATS_REQUESTS_LOCK = threading.Lock()
-
-_CUSTOMER_SDKSTATS_METRICS = None
-_CUSTOMER_SDKSTATS_METRICS_STATE_LOCK = threading.Lock()
-
-_CUSTOMER_SDKSTATS_STATE = {
-    "SHUTDOWN": False,
-}
-_CUSTOMER_SDKSTATS_STATE_LOCK = threading.Lock()
-
-_LOCAL_STORAGE_SETUP_STATE = {
-    "READONLY": False,
-    "EXCEPTION_OCCURRED": ""
-}
-
-_LOCAL_STORAGE_SETUP_STATE_LOCK = threading.Lock()
 
 def is_statsbeat_enabled():
     disabled = os.environ.get(_APPLICATIONINSIGHTS_STATSBEAT_DISABLED_ALL)
@@ -90,41 +71,6 @@ def set_statsbeat_live_metrics_feature_set():
         _STATSBEAT_STATE["LIVE_METRICS_FEATURE_SET"] = True
 
 
-def get_customer_sdkstats_shutdown():
-    return _CUSTOMER_SDKSTATS_STATE["SHUTDOWN"]
-
-
-def get_local_storage_setup_state_readonly():
-    return _LOCAL_STORAGE_SETUP_STATE["READONLY"]
-
-
-def set_local_storage_setup_state_readonly():
-    with _LOCAL_STORAGE_SETUP_STATE_LOCK:
-        _LOCAL_STORAGE_SETUP_STATE["READONLY"] = True
-
-
-def get_local_storage_setup_state_exception():
-    return _LOCAL_STORAGE_SETUP_STATE["EXCEPTION_OCCURRED"]
-
-
-def set_local_storage_setup_state_exception(value):
-    with _LOCAL_STORAGE_SETUP_STATE_LOCK:
-        _LOCAL_STORAGE_SETUP_STATE["EXCEPTION_OCCURRED"] = value
-
 def set_statsbeat_shutdown(shutdown: bool):
     with _STATSBEAT_STATE_LOCK:
         _STATSBEAT_STATE["SHUTDOWN"] = shutdown
-
-
-def is_customer_sdkstats_enabled():
-    return os.environ.get(_APPLICATIONINSIGHTS_SDKSTATS_ENABLED_PREVIEW, "").lower() == "true"
-
-
-def get_customer_sdkstats_metrics():
-    return _CUSTOMER_SDKSTATS_METRICS
-
-# pylint: disable=global-statement
-def set_customer_sdkstats_metrics(metrics):
-    global _CUSTOMER_SDKSTATS_METRICS
-    with _CUSTOMER_SDKSTATS_METRICS_STATE_LOCK:
-        _CUSTOMER_SDKSTATS_METRICS = metrics
